Closes #4251: Use GitHub Actions matrix for PHPUnit + optimize existing tests for concurrency. (#4252)
* Closes #4251: Use GitHub Actions matrix feature for PHPUnit test concurrency.


---------
