+++ modules/custom/az_demo/tests/src/Functional/AZDemoContentTest.php
@@ -55,34 +55,22 @@ class AZDemoContentTest extends QuickstartFunctionalTestBase {
   }
 
   /**
-   * Tests that imported utility links exist.
+   * Tests demo content.
    */
-  public function testUtilityLinks() {
+  public function testDemoContent() {
     // Go to the front page.
     $this->drupalGet(Url::fromRoute('<front>'));
     $assert = $this->assertSession();
     $assert->statusCodeEquals(200);
+
     // Check for utility links.
     $assert->linkExists('Utility 1');
     $assert->linkExists('Utility 2');
-  }
-
-  /**
-   * Tests page titles.
-   */
-  public function testTitle() {
-    $this->drupalGet(Url::fromRoute('<front>'));
-    $assert = $this->assertSession();
-    $assert->statusCodeEquals(200);
 
     // Home page title test.
     $assert->elementContains('css', '#block-az-barrio-page-title h1.title span.field--name-title', 'Kitten');
-  }
 
-  /**
-   * Tests publication links.
-   */
-  public function testPublicationLinks() {
+    // Tests publication links.
     $this->drupalGet('/publications');
     $assert = $this->assertSession();
     // Assert the page loads successfully.
