+++ themes/custom/az_barrio/tests/src/Functional/AzBarrioTest.php
@@ -31,87 +31,21 @@ class AzBarrioTest extends QuickstartFunctionalTestBase {
   protected $strictConfigSchema = FALSE;
 
   /**
-   * The created user.
-   *
-   * @var \Drupal\user\Entity\User
+   * Test AZ Barrio as an anonymous user.
    */
-  protected $adminUser;
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function setUp(): void {
-    parent::setUp();
-
-    // Create a test user.
-    $this->adminUser = $this->drupalCreateUser([
-      'access administration pages',
-      'administer themes',
-      'administer blocks',
-    ]);
-    $this->drupalLogin($this->adminUser);
-  }
-
-  /**
-   * Test AZ Barrio's defaults.
-   */
-  public function testThemeDefaults() {
+  public function testAnonymous() {
+    // Test AZ Barrio's defaults.
     $this->drupalGet('');
     $this->assertSession()->statusCodeEquals(200);
     $this->assertSession()->responseContains('https://fonts.googleapis.com/css?family=Material+Icons+Sharp');
     $this->assertSession()->responseContains('https://use.typekit.net/emv3zbo.css');
-  }
 
-  /**
-   * Tests that the Arizona Barrio theme can be uninstalled.
-   */
-  public function testIsUninstallable() {
-    // Login.
-    $this->drupalLogin($this->adminUser);
-    $this->drupalGet('admin/appearance');
-    $this->cssSelect('a[title="Set Bootstrap Barrio as default theme"]')[0]->click();
-    $this->cssSelect('a[title="Uninstall Arizona Barrio theme"]')[0]->click();
-    $this->assertSession()->pageTextContains('The Arizona Barrio theme has been uninstalled.');
-  }
-
-  /**
-   * Tests that the header column class settings work on install.
-   */
-  public function testHeaderColumnClassesAreSet() {
+    // Tests that the header column class settings work on install.
     $this->drupalGet('');
     $this->assertSession()->elementExists('css', '#header_site > div:nth-child(1) > div > div.col-12.col-sm-6.col-lg-4');
     $this->assertSession()->elementExists('css', '#header_site > div:nth-child(1) > div > div.col-12.col-sm-6.col-lg-8');
-  }
 
-  /**
-   * Tests that the navbar off-canvas region classes are set on install.
-   */
-  public function testNavbarOffCanvaseRegionClassesSetonInstall() {
-    $this->drupalGet('');
-    $this->assertSession()->elementExists('css', '#navbar-top.navbar-offcanvas.has-navigation-region.has-off-canvas-region');
-  }
-
-  /**
-   * Tests that the navbar off-canvas region classes are set properly.
-   *
-   * When blocks are removed or added to regions, classes should change.
-   */
-  public function testNavbarOffCanvaseRegionClassesSetNoOffCanvasBlocks() {
-    $this->drupalLogin($this->adminUser);
-    $this->drupalGet('admin/structure/block');
-    $this->cssSelect('ul[data-drupal-selector="edit-blocks-az-barrio-offcanvas-searchform-operations"] li.disable a')[0]->click();
-    $this->drupalGet('');
-    $this->assertSession()->elementExists('css', '#navbar-top.navbar-offcanvas.has-navigation-region.no-off-canvas-region');
-    $this->drupalGet('admin/structure/block');
-    $this->cssSelect('ul[data-drupal-selector="edit-blocks-az-barrio-main-menu-operations"] li.disable a')[0]->click();
-    $this->drupalGet('');
-    $this->assertSession()->elementExists('css', '#navbar-top.navbar-offcanvas.no-navigation-region.no-off-canvas-region');
-    $this->drupalGet('admin/structure/block');
-    $this->cssSelect('ul[data-drupal-selector="edit-blocks-az-barrio-offcanvas-searchform-operations"] li.enable a')[0]->click();
-    $this->drupalGet('');
-    $this->assertSession()->elementExists('css', '#navbar-top.navbar-offcanvas.no-navigation-region.has-off-canvas-region');
-    $this->drupalGet('admin/structure/block');
-    $this->cssSelect('ul[data-drupal-selector="edit-blocks-az-barrio-main-menu-operations"] li.enable a')[0]->click();
+    // Tests that the navbar off-canvas region classes are set on install.
     $this->drupalGet('');
     $this->assertSession()->elementExists('css', '#navbar-top.navbar-offcanvas.has-navigation-region.has-off-canvas-region');
   }
